feat(#S11P21D109-284): 거래내역 구현

+++ backend/spring-server/src/main/java/com/semonemo/spring_server/domain/coin/dto/request/CoinServiceRequestDto.java
@@ -13,5 +13,5 @@ import java.math.BigInteger;
 @ToString
 public class CoinServiceRequestDto {
     private Long userId;
-    private BigInteger amount;
+    private Long amount;
 }
